requirements.txt
Flask==3.0.3
Werkzeug==2.3.2
mysql-connector-python==8.0.33
PyJWT==2.8.0
pytz==2023.3
python-dotenv==1.0.0